+++ src/components/Deletions/undelete/createTpopmassn.js
@@ -11,7 +11,7 @@ export default gql`
     $datum: Date
     $bemerkungen: String
     $planBezeichnung: String
-    $flaeche: Int
+    $flaeche: Float
     $markierung: String
     $anzTriebe: Int
     $anzPflanzen: Int
