+++ src/components/Projekte/Daten/Tpopmassn/index.js
@@ -56,7 +56,7 @@ const fieldTypes = {
   datum: 'Date',
   bemerkungen: 'String',
   planBezeichnung: 'String',
-  flaeche: 'Int',
+  flaeche: 'Float',
   markierung: 'String',
   anzTriebe: 'Int',
   anzPflanzen: 'Int',
